# BACKEND/contracts/openapi/gamification.json
{
    "openapi":  "3.1.1",
    "info":  {
                 "title":  "FraudCell.Gamification.Service | v1",
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://127.0.0.1:8083/"
                    }
                ],
    "paths":  {
                  "/api/v1/game/profiles/me":  {
                                                   "get":  {
                                                               "tags":  [
                                                                            "Gamification"
                                                                        ],
                                                               "operationId":  "GetMyProfile",
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponseOfProfileResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "400":  {
                                                                                             "description":  "Bad Request",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "401":  {
                                                                                             "description":  "Unauthorized",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "403":  {
                                                                                             "description":  "Forbidden",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "404":  {
                                                                                             "description":  "Not Found",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "409":  {
                                                                                             "description":  "Conflict",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "429":  {
                                                                                             "description":  "Too Many Requests",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "500":  {
                                                                                             "description":  "Internal Server Error",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         },
                                                                                 "503":  {
                                                                                             "description":  "Service Unavailable",
                                                                                             "content":  {
                                                                                                             "application/json":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/ApiResponse"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                         }
                                                                                         }
                                                                             }
                                                           }
                                               },
...
              },
    "components":  {
                       "schemas":  {
                                       "ApiError":  {
                                                        "required":  [
                                                                         "code",
                                                                         "message"
                                                                     ],
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "code":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                           "message":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                           "details":  {
                                                                                           "type":  [
                                                                                                        "null",
                                                                                                        "object"
                                                                                                    ]
                                                                                       }
                                                                       }
                                                    },
...
                                       "ApiResponse":  {
                                                           "required":  [
                                                                            "success",
                                                                            "meta"
                                                                        ],
                                                           "type":  "object",
                                                           "properties":  {
                                                                              "success":  {
                                                                                              "type":  "boolean"
                                                                                          },
                                                                              "data":  {

                                                                                       },
                                                                              "error":  {
                                                                                            "oneOf":  [
                                                                                                          {
                                                                                                              "type":  "null"
                                                                                                          },
                                                                                                          {
                                                                                                              "$ref":  "#/components/schemas/ApiError"
                                                                                                          }
                                                                                                      ]
                                                                                        },
                                                                              "meta":  {
                                                                                           "$ref":  "#/components/schemas/ApiMeta"
                                                                                       }
                                                                          }
                                                       },
...
                                       "ApiResponseOfProfileResponse":  {
                                                                            "required":  [
                                                                                             "success",
                                                                                             "meta"
                                                                                         ],
                                                                            "type":  "object",
                                                                            "properties":  {
                                                                                               "success":  {
                                                                                                               "type":  "boolean"
                                                                                                           },
                                                                                               "data":  {
                                                                                                            "oneOf":  [
                                                                                                                          {
                                                                                                                              "type":  "null"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "$ref":  "#/components/schemas/ProfileResponse"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                               "error":  {
                                                                                                             "oneOf":  [
                                                                                                                           {
                                                                                                                               "type":  "null"
                                                                                                                           },
                                                                                                                           {
                                                                                                                               "$ref":  "#/components/schemas/ApiError"
                                                                                                                           }
                                                                                                                       ]
                                                                                                         },
                                                                                               "meta":  {
                                                                                                            "$ref":  "#/components/schemas/ApiMeta"
                                                                                                        }
                                                                                           }
                                                                        },
...
                                       "ProfileResponse":  {
                                                               "required":  [
                                                                                "analystId",
                                                                                "displayName",
                                                                                "totalPoints",
                                                                                "level",
                                                                                "dailyRank",
                                                                                "weeklyRank",
                                                                                "totalDecisions",
                                                                                "averageDecisionSeconds",
                                                                                "accuracyRate",
                                                                                "badges"
                                                                            ],
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "analystId":  {
                                                                                                    "type":  "string"
                                                                                                },
                                                                                  "displayName":  {
                                                                                                      "type":  [
                                                                                                                   "null",
                                                                                                                   "string"
                                                                                                               ]
                                                                                                  },
                                                                                  "totalPoints":  {
                                                                                                      "type":  "integer",
                                                                                                      "format":  "int64"
                                                                                                  },
                                                                                  "level":  {
                                                                                                "type":  "string"
                                                                                            },
                                                                                  "dailyRank":  {
                                                                                                    "type":  [
                                                                                                                 "null",
                                                                                                                 "integer"
                                                                                                             ],
                                                                                                    "format":  "int32"
                                                                                                },
                                                                                  "weeklyRank":  {
                                                                                                     "type":  [
                                                                                                                  "null",
                                                                                                                  "integer"
                                                                                                              ],
                                                                                                     "format":  "int32"
                                                                                                 },
                                                                                  "totalDecisions":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int32"
                                                                                                     },
                                                                                  "averageDecisionSeconds":  {
                                                                                                                 "type":  [
                                                                                                                              "null",
                                                                                                                              "number"
                                                                                                                          ],
                                                                                                                 "format":  "double"
                                                                                                             },
                                                                                  "accuracyRate":  {
                                                                                                       "type":  "number",
                                                                                                       "format":  "double"
                                                                                                   },
                                                                                  "badges":  {
                                                                                                 "type":  "array",
                                                                                                 "items":  {
                                                                                                               "$ref":  "#/components/schemas/BadgeSummaryResponse"
                                                                                                           }
                                                                                             }
                                                                              }
                                                           }
                                   }
                   },
...
}